+++ tests/diff_fixture/new/added.py
@@ -1,0 +1,3 @@
+class New:
+    def method(self):
+        pass