+++ tests/diff_fixture/old/mod.py
@@ -1,0 +1,9 @@
+import a
+
+
+def f(x):
+    pass
+
+
+def g():
+    pass